requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
idna==2.10
mccabe==0.6.1
pycodestyle==2.7.0
pyTelegramBotAPI==3.7.9
python-dotenv==0.17.1
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.5
